Pin `six` python package version to ~>1.15
This commit updates the build spec in order to provide an explicit
version range for `six` pip package to ~>1.15 (>=1.15.0, <1.16.0).

The current pipeline comes with pip version 1.14.0, which is incompatible
with the new cfn-cli tool and breaks the build pipeline.

Providing a patch-level version range allows the pipeline to receive
patch versions with no need to change the spec again.

The version range must be bumped again upon an incompatible update
landing on cfn-cli side.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>