Fix OptionGroup contract tests
This commit introduces minor changes in OptionGroup contract test
fixtures. OptionSettings are not being provided on object creation and
update.

The reason is: ModifyOptionGroup RDS API call changes the order
of the OptionSettings list and CFN CLI tool expects all collections to
be preserved in exactly the same order as they are provided in the input
fixtures even if `insertionOrder=false` is set. This behavior is
confirmed to be unexpected and the corresponding corrective action is on
CFN CLI team backlog.

For the time being, we stick to providing no OptionSettings in the test
fixtures.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>