Implement AWS::RDS::DBInstance error rulesets (#69)
This commit introduces a new mechanism for RDS API response error handling in
AWS::RDS::DBInstance resource.

The motivation for this change is to provide a contextual error handling instead
of the existing static 1-size-fits-all approach. In other words, the existing
approach handles API error codes the same way for all resource handlers (create,
update, delete), which can cause an unwanted effect.

Another reason for this change is to handle generic AwsServiceException class of
errors. RDS SDK exposes a limited range of dedicated exception classes, the rest
is reported using a generic AwsServiceException error with AwsErrorDetails
container attribute including the error code.

The new approach introduces the concept of error rule sets. A rule set is
supposed to consume an exception instance and output a corresponding error
handling status.

There are 3 statuses supported by the error sets:
* `IgnoreErrorStatus`, ignore
* `HandlerErrorStatus`, fail with a specific handler error code
* `UnexpectedErrorStatus`, fail with an unexpected exception

This status instructs BaseHandler proceed as there was no exception.

This status instructs BaseHandler to return a failure event with a particular
pre-defined `HandlerErrorCode` value. This is an expected failure code, hence it
is used by CFN to facilitate the resource lifecycle.

This status instructs BaseHandler to return an unexpected failure code (internal
failure code) back to the invoker. This is the last resort status that is returned
by any error rule set if there is no corresponding handling rule defined for
this error class and/or error code.

Error rule sets are defined per handler and are falling back to the default rule
set that is aiming to handle generic issues like auth, throttling or client
errors.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>