Fix AWS::RDS::OptionGroup on empty option configurations (#70)
This commit changes OptionGroup Create handler behavior on a case when the
option configurations collection is empty. The old behavior used to call
UpdateOptionGroup anyways, which resulted in a 400 response from the API
backend. This commits turns this update into a conditional operation by checking
if the underlying collection of option configurations is non-empty.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>