Fix AWS::RDS::DBInstance validation schema
This commit introdues a series of changes that are aiming to restore the
backwards compatibility in the resource implementation.

In particular:
- Validation pattern accept empty strings as empty values for:
  - `DBClusterIdentifier`
  - `DBInstanceIdentifier`
  - `DBName`
  - `DBParameterGroupName`
  - `DBSnapshotIdentifier`
  - `DBSubnetGroupName`
- Removed strict enum-based `Engine` type validation: using an enum
enforces a case sensitive validation and breaks existing clients.
- `Port` is marked as write-only because RDS API will ensure to populate
the `Endpoint` attribute, and `Port` update depends on the current state
of the instance.
- Resource identifier generation performs a secondary pass replacing
double dashes with a singular character.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>