Set default MultiAZ flag on AWS::RDS::DBInstance restore from snapshot
This commit changes the default MultiAZ flag behavior on restoring a db
instance from a snapshot. The previous behavior was consuming the flag from the
request object with no change.

In fact, the default value for MultiAZ flag is engine-dependant. Whereas for the
majority of the engines the default walue would be set to false, SQLServer
engines supporting mirroring expect this value set to null upon a restore.

A new IAM policy rds:DescribeDBSnapshots has been added to Create handler
permissions.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>