Check resource immutability on AWS::RDS::DBInstance update
This commit adds an explicit DBInstance resource immutability check upon an
update on top of the schema-based CreateOnly attribute set.

The motivation for this change is to ensure the update can be updated to the
desired state. Otherwise a colmplete replacement is required.

The check logic is consolidated in a new utility class called
`ImmutabilityHelper`. The check is performed on a set of attributes:
  * Engine
  * MultiAZ
  * PerformanceInsights

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>