Re-implement AWS::RDS::DBCluster
This commit contains a deep refactoring of AWS::RDS::DBCluster handler
resource.

The changes are progressively building on top of
https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/pull/81

The refactoring is touching all action handlers and unit tests. The
package imports aws-rds-cfn-common and uses it broadly in the test suite
and error handling.

ListTagsForResource was dropped from both: permission list and factual
use in the handlers. This implementation relies on
DescribeDBClustersResult containing the tag list.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>