Ignore InvalidDBInstanceStateFault if DBInstance when being deleted (#88)
This commit changes AWS::RDS::DBInstance DeleteHandler behavior.

The current implementation interrupts the execution chain and returns a
resource conflict error code back. This might lead to a conflicting CFN
stack state where an out-of-bounds delete request blocks the stack
removal.

With the new behavior the handler will ignore the exception if the
instance is being deleted and go straight to the stabilization loop. In
this case it doesn't matter that an instance was modified from outside,
the handler will simply wait until the resource is gone and complete the
modification cycle.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>