Unset DBInstance MaxAllocatedStorage
This commit addresses unsetting `MaxAllocatedStorage` property upon a
DBInstance update.

Given a CFN template (state A):

  Resources:
    DBInstance:
      Type: AWS::RDS::DBInstance
      Properties:
        AllocatedStorage: 100
        ...

Provided an update (state B):

  Resources:
    DBInstance:
      Type: AWS::RDS::DBInstance
      Properties:
        AllocatedStorage: 100
        MaxAllocatedStorage: 1000
        ...

Transition to state B will set `MaxAllocatedStorage` to 1000 and enable
the autoscaling.

If `MaxAllowedStorage` is unset later in the template (transition back to
state A or its equivalent), the current implementation will keep the property
in place rather than unsetting it.

RDS API does not have an explicit request parameter to unset this property.
The way to get around it is to set `MaxAllocatedStorage` to the current
`AllocatedStorage`.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>