Handle RDS InternalFailure
This commit adds a new ErrorCode called `InternalFailure`. It is
returned by RDS API in a case of a service event.

An additional change to DBInstance and DBInstance role stabilization
logic: errors are handled by `Commons.handleException/3` instead of the
custom local try-catch block because it fails with a stack trace and an
internal error status if RDS API returns a 5xx response code.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>