[DBInstance] Fix never-stabilizing delete requests
This PR addresses an issue with `DeleteHandler` where
`DBSnapshotIdentifier` resource attribute is being copied to
`FinalDBSnapshotIdentifier` `delete-db-instance` request attribute. The
validation logic for these 2 attributes on RDS API side differs: whereas
the former can be a fully-qualified ARN, the latter must be a plain
identifier. The identifier copy part turned out to be inconsistent with
the legacy implementation, hence this PR enforces a snapshot id
re-generation upon a delete call.

This PR also changes the original behavior of `DeleteHandler` which used to
ignore `InvalidParameterValue` error code. The fix changes this to
throwing a terminal error so the customer can address the validation
issue.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>