Handle Error: Invalid Client Token Id
This PR adds an explicit error handler for InvalidClientTokenId error
code.

AWS RDS Service might eventually return this error code if the
credential token is stale. There is no better way to handle this rather
than re-try in a bit: the error seems to be transient.

This PR treats InvalidClientTokenId as an internal service error.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>