[DBCluster] Set Engine and MasterUserPassword conditionally
This PR changes the way ModifyDBCluster request attributes are being
set. In particular, Engine and MasterUserPassword are being set
conditionally.

Engine should only be set if the previous and the desired values differ
and the stack is not undergoing a rollback. Otherwise RDS API will
return an error "Cannot modify engine version without a healthy primary
instance in DB cluster".

MasterUserPassword should only be set if the values differ.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>