[Common] Commons handler treats IgnoreErrorStatus as success (#106)
This PR changes the way Commons handler interprets IgnoreErrorStatus
upon an exception handling. The current implementation always handles
this code as an in-progress event. This cr makes it configurable and
returns a success event by default instead.

The reason for this change is the fact that an in-progress event chain
won't execute the upcoming `then` stage right away, but will spin up
another event processing cycle. This creates an unwanted processing
cycle in cases like DBInstance when associated roles are being added.

On the opposite side, a success-event could be a subject for an early
return from the chain, even if the chain is intentionally terminated
with `progress()` invoker. For this purpose IgnoreErrorStatus is made
configurable so the end invoker can declare the desired event status.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>