[DBInstance] Handle drifting update requests
This code change adds an alternative Update handler branch to facilitate
a DBInstance resource drift resolution.

The change checks the request `Driftable` flag and performs a
scatter-status check for the supporting objects. The check validates
`in-sync` status for the following objects:
  - DBParameterGroup
  - OptionGroup
  - DBClusterParameterGroup (if applicable)

Upon a complete sync, the handler would check if the instance requires a
restart and performs a `reboot-await`.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>