[DBCluster] Force final snapshot ID generation upon a deletion
This code change enforces a final snapshot ID generate if a snapshot is
requested. In contrast with the current implementation, where the ID
generation was conditional (`DBSnapshotIdentifier` was reused), the new
behavior generates a unique ID when a resource is being deleted.

The motivation for this change is to eliminate a customer snapshot
override. This strategy was added by default and is seen wrong now.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>