[DBInstance] Fix DB instance is already in the target DB subnet group
This commit removes setting DBSubnetGroup from ModifyDBInstance requests
in order to delegate it to
CreateDBInstance/RestoreDBInstanceFromSnapshot/CreateDBInstanceReadReplica
requests as `DBSubnetGroupName` is marked as a create-only property in
the resource schema. This would guarantee that CFN mechanisms would
engage once the attribute value is altered.

At the moment an attempt to restore db instance from a snapshot would
fail with a 400 error code from RDS API because the target subnet group
would be provided with both restore- and modify-requests.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>