[DBInstance] Soft fail stack-level tag updates (#113)
This code change introduces a soft failing mechanism for stack-level
tags upon a DBInstance update/create handler invocation.

The motivation for this change is to keep a complete backwards
compatibility with the existing implementation. The point of interest is
whether a customer has the corresponding IAM permissions:
  - rds:AddTagsToResource
  - rds:RemoveTagsFromResource

The current logic is:
  - System tags are created and managed by CloudFormation. RDS API
  recognizes these tags and requires no extra IAM permissions to create
  these tags.
  - Resource-level tags can only be provided by the customer, hence
  require both permissions mentioned above.
  - Stack-level tags are declared by the customer but replicated to the
  resources by CloudFormation. In the light of the fact that distinct
  resources handle stack tags AccessDenied error in a different way, RDS
  tends to act the "soft way" and ignores these errors.

This change came with a 2-stage tag create: the initial create/restore
operation only creates system-level tags. The next stage is the
resource- and stack-level tag update, which is packed in a single
request. If the request contains no resource-level tags (stack-level
only), an AccessDenied error would be soft-handled. In case there are
customer-provided resource tags, the request would hard fail.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>