[DBInstance] Fix unstable collection order
This PR addresses a problem of an unstable ordering in the request
objects created by Translator and supplementary midleware. The essence
of the problem is in the hashCode computed for a particular request by
the CFN core library. A request hashCode is used as a part of the
callback context call graph key and is expected to be invariant in
between multiple invocations of the same call chain. If a collection
order is unstable, it can change the hash entrophy and result to a
distinct hashCode sum. This causes CFN to loose track of the previous
progress and restart it from the beginning.

On practice it means that an unstable request object hash code could
cause CFN to re-run an already executed stage twice, causing the service
to terminate the stack with an error.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>