[Common] Register Jackson JavaTimeModule explicitly (#128)
This PR changes introduces an explicit import of Jackson JavaTimeModule
library and registers it in the FilteredJsonPrinter ObjectMapper.

The reason for having this import and registration is because the module
autoscan does not work under some build systems, e.g. Groovy. Maven
picks this up just fine, but it is a non-generic approach. This code
change ensures the import is enforced.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>