[DBInstance] Fix Port data type from integer to string (#129)
This PR changes `port` attribute data type from integer to string. It
seems the old implementation was using a string attribute, and the new
implementation translated it into an integer. The fix is required as
customers who were using string functions (e.g. `Fn::Join`) experienced
a regression due to an argument of an invalid data type provided.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>