[DBInstance] Disable engine setting upon an update
This commit removes an engine set enforcement from the UpdateHandler.
The reason be: `ModifyDBInstance` never updates the engine name itself:
it only checks if there is an engine name change and triggers a
replacement cycle if this is the case.

Setting an engine in this part of the code caused a regression under
which the handler invoked `ModifyDBInstance` request once and failed
upon a stabilization right away as `ensureEngineSet` invocation updated
the desired resource model by reference. A close review of this code
indicates that setting the engine name here is a no-op, hence it is
removed.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>