[DBInstance] Increase default handler request chain timeout (#151)
This commit updates the default request chain timeout from 60 to 90
minutes. It seems some db instances time out to stabilize within 1hr if
it observes a stabilization interruption. The stabilization for
DBInstance is n-probed: the handler expects to observe 3 `available`
statuses before it declares a stable state. If the counter is reset due
to an interrupting instance reboot or upgrade, the handler will reset
the probe counter and start it over. We are extending this default
timeout in order to cover slow-provisioning cases.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>