[DBInstance] Remove `StorageType` enum-based validator
This commit removes the enum-based validator that was defined on
`StorageType` attribute previously. The reason for the removal is
because CFN performs a case-sensitive enum lookup, so the values in
different case spelling would be rejected.

A case-insensitive enum would rather be an option on CFN side, hence
this commit could be rejected once ignore-case check is supported by CFN.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>