[DBInstance] Increase MasterUsername maxlen constraint
This commit alters the maxlen constraint for `MasterUsername` from 16
characters to 128 max. The reason for change is that initially
DBInstance schema was aiming for the minimum length across all engines,
but it causes validation failures on engines where longer usernames are
allowed.

According to
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-rds-database-instance.html#cfn-rds-dbinstance-masterusername,
the maxlen across all engines is 128 characters. If a csutomer provides
a username that is longer than the engine maxlen but below 128
characters, the value would be sent to RDS API and rejected by the
backend on contrast to rejecting by the the CFN validator. We believe
this is a reasonable tradeoff that would allow us to include all engine
constraints.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>