[DBCluster] Remove default parameter group name set on cluster create
This commit removes the code that sets the default DBCluster parameter
group name upon creation. The reason for change is that RDS API picks up
the default parameter group name automatically depending on the engine
version. The previous implementation hardcoded the group name to
aurora5.6, which caused troubles on other engine versions.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>