[Common] Introduce `IdentifierFactory`
This commit introduces a new util factory class called
`IdentifierFactory`. The aim is to provide a generic and safe approach
for new identifier generation.

The current approach adopted across all RDS CFN resources uses
cloudofrmation code IdentifierUtils, which appeared to be missing the
cases of double hyphens and is not null-safe to the identifier
components: stack id, logical resource identifier and request token.

The newly added util generalizes the identifier generation approach and
hides null-safety checks.

This commit affects all RDS CFN resources. Some areas of the code were
changed due to an auto-formatting and auto-rearranges.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>