[DBInstance] Remove Iops validator (#171)
This commit removes validation constraint for `Iops` attribute. The
reason for this change is that some engine types do not support setting
iops and customers set it to 0 to keep it consistent across all the
templates in use. RDS API treats a 0-value as null, hence it is a valid
parameter.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>