[DBCluster] Fix wrong Assert import (#173)
This commit fixes a wrong unit test code assert import. The import was
added by mistake. The library that was meant to be imported is
`org.assertj.core.api.Assertions`. Due to a distinction in the CI/CD
pipeline environments and imported suite, the pipeline experienced
failures. This commit ensures the same Assertions library is used across
all resources.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>