[DBCluster] Fix failing AddRoleToDBCluster stabilization
This commit fixes an error in DBClusterRole comparison utility. It is
being used by the ModifyDBCluster stabilization checker.

The faulty implementation compared the inner representation of
DBClusterRole to the SDK version it received from DescribeDBClusters
call. Upon a comparison, both ARN and FeatureName from the internal
representation were safe-wrapped by Optional with a fallback to an empty
string. The same null-safe wrapping was not performed for the SDK
representation, causing an empty string being compared to null. This
comparison never succeeded, hence the handler timed out the resource
stabilization.

This commit alters the comparison technique. It uses Objects.equals
utility instead, which provides both null-safety and correctness in case
any of the arguments is null.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>