[DBInstance] Use default handler config (#176)
This commit introduces a default value for the handler config. The
motivation for the change is that every handler instantiates a custom
configuration of `HandlerConfig` with eventually inconsistent config
between the handlers. Whereas this was the initial idea, in practice it
seems to be more error-proof to declare a unified one-size-fits-all
config per resource and tweak the config attributes in a single place.

Test code is still encouraged to override the handler config
independently, depending on the use-case.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>