[DBInstance] Fix resource drift stabilization (#177)
This commit fixes DBInstance UpdateHandler behavior when it is being
invoked with `driftable=true` flag. This is a special case that occurs
when CFN detects a resource drift and executes an additional series of
stabilization loops. The entire idea behind this check is to see if the
instance is required a reboot and ensuring it completes successfully.

In the current implementation reboot and stabilization loops are placed
in a wrong order, causing the handler to wait forever upon a
`pending-reboot` status and never reaching the point of the actual
reboot. This fix changes this order and ensures the reboot is performed
before the stabilization chain is started.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>