[GlobalCluster] Fix GlobalCluster create for ARN SourceDBClusterIdentifier
This commit fixes a bug in GlobalCluster CreateHandler that is ignoring
sourceDBClusterIdentifier if it is provided as an ARN. The handler logic
made an exception for an ARN-looking identifier and avoided a cluster
fetch operation, but never specified the source identifier in the create
request.

This commit ensures the source ARN is taken into account upon a create
request.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>