[DBInstance] Versioned RDS API client (#182)
This commit introduces version-aware `VersionedProxyClient` and the
semantics of the SDK API versioning. The motivation for this change is
to facilitate legacy fleet of RDS `DBInstance` resources that declare
deprecated `DBSecurityGroups` attribute.

RDS API enforces the client to be explicit about the API version in use
and rejects the default version. If a `DBInstance` resource declares a
non-empty `DBSubnetGroups` collection, the handler will downgrade the
rdsClient to `v12`(`2012-09-17`). The downgrade is performed on the
following API invocations:
- `CreateDBInstance`
- `RestoreDBInsstanceFromSnapshot`
- `ModifyDBInstance`

As the SDK provides no direct handle to alter the API version in use, a
fairly convoluted approach is chosen: we introduce an interceptor that
alters the raw request query parameters with a corresponding `Version`
attribute. This approach can break if the SDK implementation concerning
interceptor invocation order changes. This commit adds a low-level test
case that ensures the SDK interceptor chain works the way we expect it
to.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>