[DBInstance] Avoid unnecessary rename upon a DBInstance update (#190)
This commit eliminates an unnecessary rename + reboot upon a database
instance update. The reason it happens is because
`Translator.modifyDBInstanceRequest` sets
`ModifyDBInstanceRequest.newInstanceIdentifier` unconditionally, causing
RDS to interpret this as an attempt to rename the instance. This
automatically causes a reboot. By setting this attribute conditionally
we eliminate the reboot and avoid a potential downtime.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>