[DBInstance] Ignore InvalidDBInstanceState on Delete
This commit introduces a conditional error handling in DBInstance Delete
handler. In particular, it instructs the rule set to ignore an
already-deleting instance state upon a deletion. Any other invalid state
errors are still being interpreted as a ResourceConflict error.

The motivation for this change is to avoid a CFN stack failure when it
enters a deleting state and observes a DBInstance that is already being
deleted (an out-of-bounds delete).

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>