[DBCluster] Restore original dbClusterIdentifier case spelling
This commit modifies ReadHandler behavior so it restores the original
dbClusterIdentifier case spelling. The motivation for this change is to
fix CFN side effects caused by RDS coalescing all identifiers to
lowercase. In particular, any mixed-case identified resource is a
subject for an automatic replacement as an identifier is a replace-only
modification for CFN.

This commit substitutes the observed (lowercased) identifier with the
one provided in the template (original spelling).

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>