[DBInstance] Do not set EngineVersion upon a rollback (#202)
This commit changes the way DBInstance UpdateHandler sets EngineVersion.
Engine version change paths are primarily defined for upgrades, not for
downgrades. It means that an attempt to set a lower engine version
(which is the case upon a rollback after an attempt to upgrade the
engine version) will be set for failure. This commit ensures the handler
does not alter the engine version if a rollback is requested.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>