[DBInstance] Fix repeating Create* call failures on tag softfail
This commit fixes an issue in CreateHandler. When the handler soft-fails
a Create* call containing all tags, it attempts to invoke the same
method with system tags only. If this call succeeds, it enters a
stabilization cycle. A ProgressEvent is re-entrant, meaning that any
code that is not guarded by the special ProgressEvent sentinels would
be re-executed every time a stabilizer returns an in-progress signal.
This effect had an impact on the safeAddTags method: whenever there was
a downgrade to system-tags only, the original all-tags method was still
invoked every time the system-tags only stabilizer returned an
in-progress. This commit ensures that the all-tags Create* attempt will
only fail once and won't be triggered upon a re-entrance.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>