[Common] Implement Tagging.safeCreate (#207)
This commit implements a method `Tagging.safeCreate` that is supposed to
handle tag softfailing logic. Effectively, this method was extracted
from `DBInstance.CreateHandler` implementation. The method accepts a
regular set of attributes like: proxy, progress event etc. It also
accepts an instance of `HandlerMethod` and a `Tagging.TagSet`. The
method then will attempt to invoke the `HandlerMethod` instance. If this
invocation results in an error, the `safeCreate` method will inspect the
error code. A `HandlerErrorCode.AccessDenied` would be treated as a
missing `rds:AddTagsToResource` permissions and the method will retry
the create attempt, but with system tags only. This would be the second
invocation of `HandlerMethod`. The result of the second invocation is
then rerutned as is.

This commit introduces a mixin data class called `TaggingContext` under
the commons package. The purpose of this class is to be added to a
`CallbackContext` class and exposing its availability by an explicit
`implements TaggingContext.Provider` class annotation. Any further
tagging-related flags are supposed to be located in this mixin class.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>