[Common] Implement inheritance chain in ErrorRuleSet
This commit alters ErrorRuleSet public API. In particular, the
propagation chain inintialization is made more explicit and mandatory.

From now on, every ErrorRuleSet is expected to inherit another
ErrorRuleSet. The only exception is an empty prototype singleton. All
error rule sets will have it on the tip of the inheritance chain.

API change specifics:

Previous way of building an ErrorRuleSet is:
  ErrorRuleSet oldRules = ErrorRuleSet.builder()
      .with...
      .build()
      .orElse(prototype);

The orElse API comes way too late and is optional. The new API enforces
prototype inheritance:
  ErrorRuleSet newRules = ErrorRuleSet
      .extend(prototype)
      .with...
      .build();

Also, an ErrorRuleSet can extend itself with extension rules:
  ErrorRuleSet extendedRules = baseRules.extendWith(extraRules);

This API returns a new instance of ErrorRuleSet with baseRules as the
prototype and extraRules overriding it. extraRules's inheritance chain
will be preserved completely:

Given a baseRules rule set:
  baseRules  = [ ruleSet1 ] -> [ ruleSet0 ] -> [ emptyRuleSet ]

Assume it is being extended with an additional rule set:
  extraRules = [ ruleSet2 ] -> [ emptyRuleSet ]

The resulting ruleset chain would look like:
  extendedRules = [ ruleSet2 ] -> [ ruleSet1 ] -> [ ruleSet0 ] -> [emptyRuleSet]

In other words, the extraRules rule set would be executed first and only
propagate down if none of the chain definitions catches the exception.

A no-match chain returns UnexpectedErrorStatus.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>