[DBClusterParameterGroup] Add default ruleset to DBCluster stabilization (#223)
This commit alters the error handling ruleset that is being used in
`waitForDbClustersStabilization`. The current implementation declares a
stand-alone error rule set that is missing rule declarations for
low-level issues like throttling and networking issues. This commit adds
an alternative or-branch to DB_CLUSTERS_STABILIZATION_ERROR_RULE_SET so
the aforementioned exceptions can be handled.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>

Co-authored-by: moataz-mhmd <73899425+moataz-mhmd@users.noreply.github.com>