[DBCluster] Handle SnapshotQuotaExceededFault and InvalidDbClusterSnapshotState
This commit adds 2 more exception classes to the BaseHandler default
error set. The reason for this change is an incomplete model error
coverage by the resource handler. Sometimes a db cluster snapshot can
either enter an unexpected state or the number of existing resources
might exceed the quota. This commit instructs the resource handler to
handle these 2 classes of exceptions.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>