[DBCluster] Fix overly restrictive schema validation
This commit is a follow-up on
https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/pull/148

The schema validation changes the constraint from `oneOf` to `anyOf`.
The intent of using `oneOf` was to declare a strict set of attribute
variations one can declare in a DBCluster resource, i.e. to enforce the
minimum required set of attributes and ensure there are no silently
ignored properties provided. This seems to be overly protective and
requires removing attributes from the template if these are not in use.
We are switching to `anyOf` approach even though non-functional
attributes will be swallowed.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>