[DBClusterParameterGroup] Fix premature DBCluster stabilization
This commit fixes an issue with the final stage of `applyParameters`.

Once the parameters have been applied, the handler code stabilizes on
the list of DBCluster objects that might depend on this
DBClusterParameterGroup. This stabilization is being short-circuited by
a flag used for unit testing purposes: a re-entrant chain invocation
skips the stabilization and reports an update complete status. Any
upcoming ModifyDBCluster calls on a dependant cluster object will cause
a `DB cluster isn't available for modification with status modifying`
API rejection.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>