[Multiple Resources] Drop schema business logic validators (#240)
This commit drops some validators introduced earlier. The change
includes enums, max values and specific engine types in DBCluster,
DBInstance and EventSubscription.

The motivation for this change is to get rid of an overly restrictive
validation on CFN end that might be out of sync with the RDS API
region-specific configuration. The validators that touch business
aspects of RDS turned out to be introducing more customer friction than
help.

In a case an invalid attribute value is provided, CFN will act as a
thin proxy and pass the value as-is to the RDS API. If the validation
fails, RDS API will return an InvalidRequest exception which would be
handled by the resource handler. RDS API messages seem to be more
user-friendly as CFN auto-generates validation messages from the schema
and they might seem cryptic.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>