[DBInstance] Change stabilization logic on update (#251)
This commit introduces a distinction between stabilization processes
after create and update. The change is introduced for the sake of a
complete logic parity with the historical implementation. In particular,
the post-update stabilization is:
- DBInstance status is `available`
- DBParameterGroup status is `in-sync`
- OptionGroup status is `in-sync`
- if DBInstance is a member of a DBCluster:
  - DBClusterParameterGroup is `in-sync`

Both stabilizers perform an additional assertion on DBInstance status.
The assertion code performs an early process interruption if DBInstance
enters one of the following statuses:
- failed
- incompatible-restore
- incompatible-network
- incompatible-parameters
- inaccessible-encryption-credentials

In this case the handler will throw a `CfnNotStabilizedException`.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>