[DBInstance] Drop reference-type schema validators
This commit removes schema-based validators from DBInstance resource
definition. The motivation for this change is to eliminate
inter-dependencies and validation logic inconsistency between CFN
resources. At the moment, DBInstance performs additional validations on
the reference-type properties.

Reference-type properties are effectively borrowed attributes:
DBInstance does not own these propoerties (they are effectively pointers
on external resources). With this change we are delegating the
reference-type validation to the corresponding owning resources or to
the RDS API if the resource was created out of CFN scope.

Having this change in place would allow us to consolidate the validation
logic in the owning resources and eliminate export issues caused by
legacy resources that would no longer pass the validation.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>