[DBInstance] Enforce final snapshot if not explicitly requested (#271)
This commit enforces the final snapshot creation on DBInstance delete in
case the request contains no data on whether a final snapshot creation
is requested or not. The reason for this change is the referencial value
type returned by `ResourceHandlerRequest.getSnapshotRequested()`. The
return value is a Boolean, meaning it can be yes/no/null(no data). This
commit addresses the latter case where the invocation request is neither
asking for a final snapshot nor disclaiming it.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>