[DBCluster] Enforce final snapshot if not explicitly requested (#272)
This commit enforces the final snapshot creation on DBCluster delete in case the request contains no data on whether a final snapshot creation is requested or not. The reason for this change is the referential value type returned by ResourceHandlerRequest.getSnapshotRequested(). The return value is a Boolean, meaning it can be yes/no/null(no data). This commit addresses the latter case where the invocation request is neither asking for a final snapshot nor disclaiming it.

Co-authored-by: Valentin Shirshov <vshir@amazon.com>