[DBInstance] Add ResourceHandlerRequest validation
This commit adds a request validation layer to the handler execution
chain. The motiovation is to extend the schema-based validation with a
range of more flexible and more dynamic validation methods.

The change introduces a new class `ValidatedRequest` that extends
ResourceHandlerRequest with 0 overloading: it is here to distinguish a
validated request from a raw one.

The change adds a new validation on the sourceRegion attribute. If a
provided region is incorrect, the validation will refuse such a request
well before it reaches the SDK endpoint connection initialization.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>